refacto(app): app refactoring on access control and roles to pass tests implemented

+++ src/Controller/TaskController.php
@@ -15,7 +15,6 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\Security\Core\Exception\InvalidCsrfTokenException;
 
 /**
  * Class TaskController.
@@ -188,6 +187,12 @@ class TaskController extends AbstractController
      */
     public function deleteTask(Task $task, TaskManager $taskManager): Response
     {
+        if (($task->getAuthor() !== null && $task->getAuthor() !== $this->getUser())
+            || ($task->getAuthor() === null && !$this->isGranted('ROLE_ADMIN'))
+        ) {
+            throw $this->createAccessDeniedException();
+        }
+
         $taskManager->deleteTask($task);
 
         $this->addFlash(
